+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import "./App.css";
 import Detail from "./components/Detail";
-import PokemonList from "./components/PokemonList";
+import PokemonList from "./components/List";
 
 function App() {
   return (
